+++ e2e/commands/status.e2e.js
@@ -450,10 +450,19 @@ describe('bit status command', function () {
         expect(output).to.have.string('non-existing dependency files');
         expect(output).to.have.string(`${path.normalize('bar/foo2.js')} -> ./foo1.js`);
       });
-      it.skip('should show an error indicating the mainFile was deleting when deleting the mainFile', () => {
-        helper.deleteFile('bar/index.js');
-        const output = helper.runCmd('bit status');
-        expect(output).to.have.string(''); // @todo: implement
+      describe('when mainFile is deleted', () => {
+        before(() => {
+          helper.reInitLocalScope();
+          helper.createFile('bar', 'index.js');
+          helper.createFile('bar', 'foo.js');
+          helper.addComponentWithOptions('bar/', { i: 'bar/foo' });
+          helper.deleteFile('bar/index.js');
+        });
+        it('should show an error indicating the mainFile was deleting', () => {
+          const output = helper.runCmd('bit status');
+          expect(output).to.have.string('invalid components');
+          expect(output).to.have.string('main-file was removed');
+        });
       });
     });
     describe('when all of the files were deleted', () => {